refactor: split Issue into its own data/issues app (D-017)
Issue was a first-class D-017 entity (persistent, cross-scan finding register)
but lived inside the findings app — inconsistent with the asset split
(raw assets/web_assets vs persistent asset_inventory). This gives findings the
same shape: raw Finding in findings/, persistent Issue in issues/.

- New apps.core.data.issues app: Issue model + issue_key, the /api/issues/
  router, and the finalize rollup (rollup_session_issues) all moved here.
- findings/ keeps only the raw Finding model + /api/findings/ router.
- Model move is STATE-ONLY: issues.Issue keeps db_table='findings_issue' and
  both migrations (findings 0010 delete, issues 0001 create) use
  SeparateDatabaseAndState — no DDL, no data migration (preprod's 49 Issue rows
  are preserved).
- Rewired ninja.py (issues_router <- issues.api, findings_router <- findings.api),
  pipeline.py rollup import, INSTALLED_APPS, test imports, and the docs app tables.

No API or behavior change; makemigrations --check clean; 159 tests across the
affected suites pass.
