+++ apps/core/data/findings/api.py
@@ -1,134 +1,28 @@
-"""Issues API — ``/api/issues/`` : the persistent, cross-scan finding register.
+"""Findings API — ``/api/findings/`` : raw per-scan Finding instances.
 
-Reads and triages the ``Issue`` register (``apps/core/data/findings/models.Issue``).
-A status change here **persists across scans** (unlike per-scan
-``Finding.status``) — that is the point of the finding-centric UI: dismiss a false
-positive once and it stays dismissed. Spec:
-docs/specs/2026-09-12-finding-centric-ui-direction.md (PR3).
+These are the per-scan ``Finding`` rows (finding-centric list + per-finding
+lifecycle status). Distinct from ``/api/issues/`` (the ``issues`` app), which is
+the deduped, cross-scan persistent register those findings are promoted into.
+Relocated here from ``/api/scans/findings/`` (D-017).
 """
 
 import logging
 import uuid
 
 from django.core.paginator import Paginator
-from django.db.models import Case, Count, IntegerField, Value, When
+from django.db.models import Case, IntegerField, When
 from django.shortcuts import get_object_or_404
 from ninja import Router, Schema
 from ninja.errors import HttpError
 
 from apps.core.console.api.auth import JWTAuth
-from apps.core.constants import SEVERITY_RANK
 
-from .models import STATUS_CHOICES, Issue
+from .models import STATUS_CHOICES
 
 logger = logging.getLogger(__name__)
 router = Router(auth=JWTAuth())
 
 _VALID_STATUSES = {s for s, _ in STATUS_CHOICES}
-# Issues that still need attention (drives the default view + the severity summary).
-_ACTIONABLE = ("open", "acknowledged", "in_progress")
-
-
-class StatusIn(Schema):
-    status: str
-
-
-def _row(i) -> dict:
-    return {
-        "id": i.id,
-        "domain": i.domain.name,
-        "source": i.source,
-        "check_type": i.check_type,
-        "title": i.title,
-        "target": i.target,
-        "severity": i.severity,
-        "status": i.status,
-        "first_seen": i.first_seen.isoformat(),
-        "last_seen": i.last_seen.isoformat(),
-        "asset_id": i.asset_id,
-    }
-
-
-def _ranked(qs):
-    """Order most-severe-first, then most-recent — using the shared SEVERITY_RANK
-    so the register's ranking can't drift from the rest of the app."""
-    whens = [When(severity=s, then=Value(r)) for s, r in SEVERITY_RANK.items()]
-    return qs.annotate(
-        sev_rank=Case(*whens, default=Value(-1), output_field=IntegerField())
-    ).order_by("-sev_rank", "-last_seen")
-
-
-@router.get("/")
-def list_issues(request, domain: str = "", status: str = "", severity: str = "",
-                source: str = "", q: str = "", page: int = 1):
-    qs = Issue.objects.select_related("domain")
-    if domain:
-        qs = qs.filter(domain__name__icontains=domain)
-    if status:
-        qs = qs.filter(status=status)
-    if severity:
-        qs = qs.filter(severity=severity)
-    if source:
-        qs = qs.filter(source=source)
-    if q:
-        qs = qs.filter(title__icontains=q)
-
-    paginator = Paginator(_ranked(qs), 25)
-    p = paginator.get_page(page)
-    return {
-        "issues": [_row(i) for i in p],
-        "total": paginator.count,
-        "page": p.number,
-        "total_pages": paginator.num_pages,
-        "has_next": p.has_next(),
-        "has_previous": p.has_previous(),
-    }
-
-
-@router.get("/summary/")
-def issues_summary(request, domain: str = ""):
-    qs = Issue.objects.all()
-    if domain:
-        qs = qs.filter(domain__name__icontains=domain)
-
-    by_status = {s: 0 for s, _ in STATUS_CHOICES}
-    for row in qs.values("status").annotate(n=Count("id")):
-        by_status[row["status"]] = row["n"]
-
-    # Severity breakdown of the still-actionable issues (open/ack/in-progress) —
-    # what actually needs work, excluding resolved / dismissed.
-    open_by_severity: dict[str, int] = {}
-    for row in (qs.filter(status__in=_ACTIONABLE)
-                  .values("severity").annotate(n=Count("id"))):
-        open_by_severity[row["severity"]] = row["n"]
-
-    return {
-        "total": qs.count(),
-        "by_status": by_status,
-        "open_by_severity": open_by_severity,
-    }
-
-
-@router.post("/{issue_id}/status/")
-def set_issue_status(request, issue_id: int, data: StatusIn):
-    if data.status not in _VALID_STATUSES:
-        raise HttpError(400, f"status must be one of {sorted(_VALID_STATUSES)}")
-    issue = get_object_or_404(Issue, id=issue_id)
-    issue.status = data.status
-    issue.save(update_fields=["status"])
-    logger.info("[issues] %s → %s", issue_id, data.status)
-    return _row(issue)
-
-
-# ---------------------------------------------------------------------------
-# Raw findings — /api/findings/ : per-scan Finding instances (finding-centric).
-# Relocated from /api/scans/findings/ (D-017: findings belong in the findings
-# namespace, not under /scans). Distinct from the /api/issues/ register above —
-# these are raw per-scan Finding rows with per-scan lifecycle status; Issues are
-# the deduped, cross-scan promotion of them.
-# ---------------------------------------------------------------------------
-
-findings_router = Router(auth=JWTAuth())
 
 
 def _serialize_finding(finding) -> dict:
@@ -161,7 +55,7 @@ class FindingStatusRequest(Schema):
     resolution_note: str | None = None
 
 
-@findings_router.get("/")
+@router.get("/")
 def list_findings(
     request,
     severity: str = "",
@@ -236,7 +130,7 @@ def list_findings(
     }
 
 
-@findings_router.post("/{finding_id}/status/")
+@router.post("/{finding_id}/status/")
 def update_finding_status(request, finding_id: int, data: FindingStatusRequest):
     from django.utils import timezone
