+++ apps/core/data/findings/models.py
@@ -135,76 +135,3 @@ class Finding(models.Model):
     def domain(self) -> str:
         """Backward-compat for DomainFinding.domain."""
         return self.target
-
-
-def issue_key(source: str, check_type: str, title: str, target: str) -> str:
-    """Stable cross-scan identity for a finding within a domain.
-
-    Keyed on (source, check_type, title, target) — the same detection on the same
-    host collapses to one persistent Issue across scans, while the same check on a
-    *different* host is a distinct issue (per-asset triage). Extends the
-    delta-detection key (source:check_type:title) with target for asset
-    granularity. Uses \\x1f (unit separator) so colons in titles/targets can't
-    collide two distinct issues into one.
-    """
-    return "\x1f".join((source or "", check_type or "", title or "", target or ""))
-
-
-class Issue(models.Model):
-    """Persistent, cross-scan identity for a finding — the *issue* that each
-    per-scan Finding is an occurrence of.
-
-    Populated by the fail-graceful rollup in ``rollup.py`` at scan finalize (never
-    by tools directly), mirroring ``asset_inventory``. Its whole reason to exist is
-    that ``status`` lives HERE, not on the per-scan ``Finding``: a triage decision
-    (``false_positive`` / ``acknowledged`` / …) therefore persists across scans
-    instead of resetting to ``open`` every run. Spec:
-    docs/specs/2026-09-12-finding-centric-ui-direction.md (PR2).
-    """
-
-    domain = models.ForeignKey(
-        "domains.Domain", on_delete=models.CASCADE, related_name="issues"
-    )
-    # Stable identity within the domain (see issue_key). Not user-facing.
-    key = models.CharField(max_length=1024)
-
-    # Denormalised display fields, refreshed from the latest occurrence.
-    source = models.CharField(max_length=50, db_index=True)
-    check_type = models.CharField(max_length=50, blank=True, db_index=True)
-    title = models.CharField(max_length=500)
-    target = models.CharField(max_length=255, blank=True)
-    severity = models.CharField(max_length=20, choices=SEVERITY_CHOICES, db_index=True)
-
-    # PERSISTENT triage status — the point of this model.
-    status = models.CharField(
-        max_length=20, choices=STATUS_CHOICES, default="open", db_index=True
-    )
-
-    first_seen = models.DateTimeField()
-    last_seen = models.DateTimeField()
-
-    # Latest occurrence + the persistent asset it sits on (grounding).
-    last_finding = models.ForeignKey(
-        "findings.Finding", on_delete=models.SET_NULL, null=True, blank=True,
-        related_name="+",
-    )
-    asset = models.ForeignKey(
-        "asset_inventory.Asset", on_delete=models.SET_NULL, null=True, blank=True,
-        related_name="issues",
-    )
-    extra = models.JSONField(default=dict, blank=True)
-
-    class Meta:
-        constraints = [
-            models.UniqueConstraint(
-                fields=["domain", "key"], name="uniq_issue_per_domain"
-            )
-        ]
-        indexes = [
-            models.Index(fields=["domain", "status", "severity"]),
-            models.Index(fields=["status", "severity"]),
-        ]
-        ordering = ["-last_seen"]
-
-    def __str__(self):
-        return f"{self.source}:{self.check_type}:{self.title[:40]} [{self.status}]"
