+++ apps/core/data/issues/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class IssuesConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "apps.core.data.issues"
+    label = "issues"
