+++ apps/core/data/issues/rollup.py
@@ -1,15 +1,12 @@
 """Roll a finished scan's findings into the persistent Issue register.
 
 Called fail-graceful from ``_finalize_session`` (after the asset rollup, so
-``Finding.asset`` links are already set). Gives each finding a cross-scan
-identity so triage status persists: a dismissed false positive stays dismissed on
-the next scan instead of resetting to ``open``. Mirrors
-``apps/core/data/asset_inventory/rollup.py``.
+``Finding.asset`` links are already set). Gives each finding a cross-scan identity
+so triage status persists: a dismissed false positive stays dismissed on the next
+scan instead of resetting to ``open``. Mirrors ``asset_inventory/rollup.py``.
 
 Idempotent (F1): re-running finalize (e.g. a DBOS step replay) converges — every
-write is an upsert with no running counters, so a replay produces the same state.
-
-Spec: docs/specs/2026-09-12-finding-centric-ui-direction.md (PR2).
+write is an upsert with no running counters. Split out of findings/ (D-017).
 """
 
 import logging
@@ -43,8 +40,6 @@ def rollup_session_issues(session) -> None:
     keys_seen: set[str] = set()
 
     with transaction.atomic():
-        # Exclude the scan_coverage meta-warning — it's a finalize-generated
-        # marker, not a real attack-surface issue (same rule as delta/count).
         for f in session.findings.select_related("asset").exclude(source="scan_coverage"):
             k = issue_key(f.source, f.check_type, f.title, f.target)
             keys_seen.add(k)
@@ -59,13 +54,10 @@ def rollup_session_issues(session) -> None:
             )
             if not created:
                 issue.last_seen = now
-                issue.severity = f.severity          # reflect the latest occurrence
+                issue.severity = f.severity
                 issue.last_finding = f
                 if f.asset_id:
                     issue.asset = f.asset
-                # Persist triage across scans; only re-open a *resolved* issue that
-                # has reappeared (regression). false_positive / acknowledged /
-                # in_progress are deliberate decisions and must survive a re-scan.
                 if issue.status == "resolved":
                     issue.status = "open"
                 issue.save(update_fields=[
